Create Loader class
Create a Loader class and ShaderLoader, TextureLoader implementations.
Provide test resources and functionality to demonstrate this functionality.

The provided demo simply renders a fixed scene - it does not use the
GameState class. It will be removed once that class is available.

The asynchronous portions of these classes should be much cleaner in
jQuery. They may eventually be rewritten.