Support animation in texture loader
Load and generate animation data in the texture loader. Associate data with
individual textures in a format easily compatible with WebGL.

Each texture frame is stored as a matrix of 32-bit floating point values.
Each animation (there may be multiple animations per texture) is stored as
a delay and an array of indices into the frames.